requirements.txt
Flask~=2.2.3
joblib~=1.2.0
numpy~=1.24.2
pandas~=1.5.3
python-dotenv~=1.0.0
requests~=2.28.2
scikit-learn~=1.2.2

sqlalchemy~=2.0.7